requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comtypes==1.1.14
cryptography==40.0.1
frozenlist==1.3.3
GPUtil==1.4.0
idna==3.4
keyboard==0.13.5
magic-filter==1.0.9
MouseInfo==0.1.3
multidict==6.0.4
numpy==1.24.2
opencv-python==4.7.0.72
pefile==2023.2.7
Pillow==9.5.0
psutil==5.9.4
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycaw==20230407
pycparser==2.21
pycryptodome==3.17
PyGetWindow==0.0.9
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.2
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
requests==2.28.2
six==1.16.0
tabulate==0.9.0
urllib3==1.26.15
Wave==0.0.2
yarl==1.8.2